Use NetworkManagerConnectionActive class
Use the NetworkManagerConnectionActive class to get the active
connection from D-Bus. This is the last item the hierarchy, and we now
have all the network information we need.

We probably want to check things like signal handling, cleanup, etc.

We also need to add logic to emit (a different signal) when a connection
changes. That way we can check against previously seen connections and
potentially display the zone selection modal.