Include TODOs about when to emit
We only actually care about changes to the connection of any wireless
devices. So we should only emit when that changes. Things like a new
wired device shouldn't result in emissions. For now these are just TODO
comments. We'll have to test when we have EventEmitter in place (ie when
running as an extension).