Make Zone Defense a GTK app
Launching the kind of window we wanted (in particular, with a dropdown)
in a GNOME Shell Extension proved challenging. We want a GTK window. But
GTK windows can't be launched in the same process as GNOME Shell, so
some other means must be used. And we would need to communicate with the
window.

So we will make the app primarily a GTK app instead. We may keep the
extension (it's hard to beat EGO as a distribution mechanism), but it
will simply launch the app.

Everything in this commit was generated by GNOME Builder.