Pass something to the window
It's ugly, and probably not the correct way. But it's something.

This shows how to do it with properties, but I'm not sure I need
properties. I can just manipulate things directly (I guess properties
give you 2-way bindings). Keeping this commit for now just as a
demonstration of how to use properties (though still not using a
binding, which is why I didn't end up needing them in the end).