Don't get zone of interface from firewalld
Getting the zone of the interface returns an empty string when it is
called (except for when the application is first started). Running
`firewall-cmd --get-zone-of-interface` even as fast as I can gives the
expected result, so this is probably a race condition. I suspect the
zone is not yet set on the interface when the method is called. We were
going to use this to show the user the current zone, as well as
pre-select the current zone in the zone picker dropdown.

This actually wasn't great anyway. If a zone was the same as the
default, you couldn't tell whether it was explicitly set to that value,
or was just set to default and would change when the default changed.

We'll have to get this information from NetworkManager instead.