Make window buttons functional
Set the zone for the connection when the user clicks 'Choose', and close
the window when the user clicks 'Exit'.

Note that right now, when the user clicks 'Choose', the window is
immediately opened again because it is triggered by changing the zone of
the connection. This should resolve itself once we begin tracking which
networks have been seen.