Save seen networks asynchronously
Note that there is a race condition here. If 2 concurrent requests are
made to add a network to seen, there is no guarantee that they will
finish in order. In fact, in my testing, the one that started first
often finished last. Meaning the the 2nd update was overwritten by the
first. If we saved synchronously, this wouldn't be a problem, because
execution would block until the save completed, forcing saves to run
serially, to completion. We could possibly put some sort of mutex in
here, but realistically, the odds of running into this in actual use are
practically zero. There is no way that after a user clicked 'Choose',
they could switch networks, and click 'Choose' again on the new window
before the previous save completed.