Don't show preferences on hamburger menu
The only thing we would ever likely have as a preference would be which
zones to display in the dropdown. I'm not sure I'd want that to be
editable via the window. And because Zone Defense can sometimes be
installed system-wide, having user specified preferences for zones would
require some sort of tiered/override logic for the values.