Handle errors reading connection-ids-seen.json
When connection-ids-seen.json can't be read, we can't reasonably
continue. However, there was no way to get that error back to
`ZoneDefenseApplication` in order to handle it. In order to do that, the
async method call to `#createConnectionIdsSeen` had to be moved out of
the constructor (and into an `async` method named `init`). Errors in
`init` are allowed to pass up to its caller
(ZoneDefenseApplication.constructor) where they can be handled. In this
case, we log the fatal error and quit.

Other minor cleanup (such as removing extraneous log statements) was
also done.