Use [prettier][1]
Mostly based on the [GJS Style Guide][2], but with some customizations.
It's a little more aggressive in making changes (ie it makes changes I
don't want - especially with `printWidth`). So it does some nice
cleanup, but requires relatively careful review of the changes. For now,
we haven't added comments to disable prettier, but we probably won't run
it very often.

[1]: https://prettier.io
[2]: https://gjs.guide/guides/gjs/style-guide.html