Clean up quit logic
Now that we're starting to call `quit` earlier, some of the cleanup it
does may be attempting to call methods on undefined variables. We'll use
optional chaining to prevent errors.

We'll also add logic to handle multiple quit requests. We end up with
multiple concurrent quit requests when signal emissions from multiple
async methods make a call to `quit`. Additionally, I suspect some of the
underlying GTK quit logic is actually multithreaded and my app actually
continues to run in the background for a little bit after calling
`super.quit`.