Make DependencyCheck emit errors
main will turn these into notifications and sometimes quit. We now emit
a non-fatal error when autostart configuration fails. We also emit fatal
errors if we can't check DBus for running services (ListNames), or
firewalld is not in the list (ie it's not running).

Having the class emit its own errors is better than creating a signal
from the caller (ie main) and passing it in to the constructor. We'll
probably also want to update NetworkState to be this way.