Use more specific name for emit error method
In ErrorSignal we had a method named emit that emitted the error. This
is fine if it's the only signal we ever emit. But in the networkState
classes, we'll also emit a signal for connection changes. So we'll use a
more specific name (emitError) for the method that emits the error.