Don't call main methods from chooseZoneWindow.js
This code was a bit circular. main.js instantiates ChooseZoneWindow. And
that instance was calling a method in main when a zone was selected.
Instead, from the ChooseZoneWindow instance, we'll emit a signal, and be
done with it. This does a better job of passing control back to main.js
where it belongs.