Only include application categories in .desktop
Application categories are in both the [`.desktop`][1] file and the
[MetaInfo][2] file. However, for the MetaInfo file, "If there’s a
launchable defined for a desktop application, categories and keywords
are pulled from the desktop file." We have a launchable defined, so
there's no need to include categories in the MetaInfo file.

We'll also fix up the categories in the `.desktop` file. We'll remove
`GTK` since the MetaInfo guidelines recommend not using it. And we'll
add the categories we already had from the MetaInfo file.

[1]: https://specifications.freedesktop.org/menu-spec/latest/category-registry.html
[2]: https://docs.flathub.org/docs/for-app-authors/metainfo-guidelines/#categories-and-keywords