Explain setup better
If a user is not connected to Wi-Fi when Bouncer is running, it won't
open a window. This is by design. But this means that when the user
first starts Bouncer to enable autostart, they may not notice anything.
To help avoid confusion, in the setup documentation we'll tell them to
expect this.