Don't include icon in MetaInfo
Commit 317c950 (Only include application categories in .desktop,
2025-01-25) removed categories from the MetaInfo file. While this is
correct, it caused flatpak-builder-lint to fail with an
`app-categories-missing` error.

Per [bbhtt][1] (a big thanks!), "app-catgeries-missing means you have a
icon tag in your metainfo drop that". Apparently, the appstream linter
takes the presence of the icon tag to indicate that it's not a desktop
application, meaning the categories tag should be present. Removing the
icon tag does prevent the linting error.

Per the [Metadata Quickstart][2], in the MetaInfo file, the launchable
"tag connects the metainfo file with a .desktop file, so AppStream
metadata generators and the distribution can absorb its metadata into
the final AppStream output." Since the .desktop file already specifies
icon, the icon tag should not be included in the MetaInfo file.

[1]: https://matrix.to/#/!xFJJgxGKItHQPfstha:matrix.org/$XAI9HqJ5DTAQxtTiYbwnWE_YQrKmXDg770aPTiUU02s?via=kde.org&via=matrix.org&via=gnome.org
[2]: https://www.freedesktop.org/software/appstream/docs/chap-Quickstart.html#qsr-app-launchable-info