Generate a notification for first run success
It can be confusing for a new user when they start Bouncer and don't see
anything happen. We already have notifications for error situations, but
not for the successful first run. So we'll add a notification to
indicate successful first run.

In order to know if it's a first run, we'll use a new file to track
whether we've already completed it.

This means that in DependencyCheck we'll have to throw an error any time
we don't want the first run success notification to be generated.